fix: DealAverageResponse 변수타입 수정

+++ BackEnd/src/main/java/com/ssafy/happy/deal/dto/DealAverageResponse.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class DealAverageResponse {
 	private int dealYear;
-	private long dealAverage;
-	private long pricePerArea;
+	private double dealAverage;
+	private double pricePerArea;
 }
